FIX: NettyStorageServer() and init() race
The code before has initialization in the constructor
where the constructor read configuration to initialize
its variables. These variables are now passed in
the init() function later on. Hence we move the code
from the constructor into a separate function which
is called later after the setting of the variables.